# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic authentication for OAuth2 client credentials",
        "title": "Easyflow OAuth2 Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@easyflow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/oauth/authorize": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Initiates the OAuth2 authorization code flow with PKCE. Requires user to be authenticated via session token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2"
                ],
                "summary": "OAuth2 Authorization endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Redirect URI (required if client has multiple registered URIs)",
                        "name": "redirect_uri",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Response type (must be 'code')",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State parameter for CSRF protection (max 255 characters)",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "PKCE code challenge",
                        "name": "code_challenge",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to redirect_uri with authorization code and state"
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - session token required",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
        "/oauth/token": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Exchange authorization code for access token, refresh tokens, or use client credentials flow",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2"
                ],
                "summary": "OAuth2 Token endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Grant type (authorization_code, client_credentials, or refresh_token)",
                        "name": "grant_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID (required if not using Basic Auth)",
                        "name": "client_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Authorization code (required for authorization_code grant)",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "PKCE code verifier (required for authorization_code grant)",
                        "name": "code_verifier",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Refresh token (required for refresh_token grant)",
                        "name": "refresh_token",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token response with access token and optional refresh token",
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_oauth.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters or grant type",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "401": {
                        "description": "Invalid client credentials",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "easyflow-oauth2-server_internal_shared_errors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code represents the HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "description": "Details contains additional error information (optional)"
                },
                "error": {
                    "description": "Error represents a predefined error code from the enum package",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.ErrorCode"
                        }
                    ],
                    "example": "INVALID_REQUEST_BODY"
                }
            }
        },
        "easyflow-oauth2-server_internal_shared_errors.ErrorCode": {
            "type": "string",
            "enum": [
                "UNAUTHORIZED",
                "NOT_ALLOWED",
                "NOT_FOUND",
                "ALREADY_EXISTS",
                "INTERNAL_SERVER_ERROR",
                "MISSING_SESSION_TOKEN",
                "INVALID_SESSION_TOKEN",
                "MISSING_CLIENT_ID",
                "INVALID_CLIENT_ID",
                "MISSING_CLIENT_SECRET",
                "INVALID_CLIENT_SECRET",
                "MISSING_CODE_CHALLENGE",
                "MISSING_STATE",
                "INVALID_STATE",
                "MISSING_REDIRECT_URI",
                "INVALID_REDIRECT_URI",
                "INVALID_CONTENT_TYPE",
                "INVALID_REQUEST_BODY",
                "MISSING_GRANT_TYPE",
                "INVALID_GRANT_TYPE",
                "MISSING_CODE",
                "INVALID_CODE",
                "MISSING_CODE_VERIFIER",
                "INVALID_CODE_VERIFIER",
                "MISSING_REFRESH_TOKEN",
                "INVALID_REFRESH_TOKEN"
            ],
            "x-enum-varnames": [
                "Unauthorized",
                "NotAllowed",
                "NotFound",
                "AlreadyExists",
                "InternalServerError",
                "MissingSessionToken",
                "InvalidSessionToken",
                "MissingClientID",
                "InvalidClientID",
                "MissingClientSecret",
                "InvalidClientSecret",
                "MissingCodeChallenge",
                "MissingState",
                "InvalidState",
                "MissingRedirectURI",
                "InvalidRedirectURI",
                "InvalidContentType",
                "InvalidRequestBody",
                "MissingGrantType",
                "InvalidGrantType",
                "MissingCode",
                "InvalidCode",
                "MissingCodeVerifier",
                "InvalidCodeVerifier",
                "MissingRefreshToken",
                "InvalidRefreshToken"
            ]
        },
...
        "internal_server_routes_oauth.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "OAuth2 access token",
                    "type": "string",
                    "example": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "description": "Lifetime in seconds of the access token",
                    "type": "integer",
                    "example": 3600
                },
                "refresh_token": {
                    "description": "OAuth2 refresh token (optional)",
                    "type": "string",
                    "example": "eyJhbGciOiJFZERTQSIsInR5cCI6..."
                },
                "refresh_token_expires_in": {
                    "description": "Lifetime in seconds of the refresh token (optional)",
                    "type": "integer",
                    "example": 86400
                },
                "scopes": {
                    "description": "Granted scopes",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "read",
                        "write"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "SessionToken": {
            "description": "Session token for authenticated users",
            "type": "apiKey",
            "name": "session_token",
            "in": "cookie"
        }
    }
}