# internal/server/routes/oauth/controller.go
// Package oauth implements the OAuth2 endpoints.
package oauth

import (
	"crypto/ed25519"
	"easyflow-oauth2-server/internal/server/middleware"
	"easyflow-oauth2-server/internal/shared/endpoint"
	"easyflow-oauth2-server/internal/shared/errors"
	"easyflow-oauth2-server/pkg/database"
	"easyflow-oauth2-server/pkg/tokens"
	"net/http"
	"net/url"
	"slices"

	"github.com/gin-gonic/gin"
	"go.uber.org/fx"
)

// Controller handles OAuth2 HTTP requests.
type Controller struct {
	service *Service
	key     *ed25519.PrivateKey
}

// ControllerParams holds dependencies for OAuthController.
type ControllerParams struct {
	fx.In
	Service *Service
	Key     *ed25519.PrivateKey
}

// NewOAuthController creates a new instance of OAuthController.
func NewOAuthController(params ControllerParams) *Controller {
	return &Controller{
		service: params.Service,
		key:     params.Key,
	}
}

// RegisterRoutes sets up the OAuth2-related endpoints.
func (ctrl *Controller) RegisterRoutes(r *gin.RouterGroup) {
	r.GET(
		"/authorize",
		middleware.SessionTokenMiddleware(ctrl.service.Config, ctrl.key),
		ctrl.Authorize,
	)
	r.POST("/token", ctrl.Token)
}

// Authorize handles the OAuth2 authorization endpoint.
// @Summary OAuth2 Authorization endpoint
// @Description Initiates the OAuth2 authorization code flow with PKCE. Requires user to be authenticated via session token.
// @Tags OAuth2
// @Accept json
// @Produce json
// @Security SessionToken
// @Param client_id query string true "Client ID"
// @Param redirect_uri query string false "Redirect URI (required if client has multiple registered URIs)"
// @Param response_type query string true "Response type (must be 'code')"
// @Param state query string true "State parameter for CSRF protection (max 255 characters)"
// @Param code_challenge query string true "PKCE code challenge"
// @Success 302 "Redirects to redirect_uri with authorization code and state"
// @Failure 400 {object} errors.APIError "Invalid request parameters"
// @Failure 401 {object} errors.APIError "Unauthorized - session token required"
// @Failure 500 {object} errors.APIError "Internal server error"
// @Router /oauth/authorize [get].
func (ctrl *Controller) Authorize(c *gin.Context) {
	utils, errs := endpoint.SetupEndpoint[any](c, endpoint.WithoutBody(), endpoint.WithUser())
	if len(errs) > 0 {
		endpoint.SendSetupErrorResponse(c, errs)
		return
	}

	// Extract client_id first to get client info for redirect_uri validation
	clientID, ok := c.GetQuery("client_id")
	if !ok || clientID == "" {
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.MissingClientID,
			"The client_id query parameter is required",
		)
		return
	}

	client, err := ctrl.service.GetClient(c.Request.Context(), clientID, c.ClientIP())
	if err != nil {
		c.JSON(err.Code, err)
		return
	}

	// Extract and validate redirect_uri early
	redirectURI, ok := c.GetQuery("redirect_uri")
	if (!ok || redirectURI == "") && len(client.RedirectUris) > 1 {
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.MissingRedirectURI,
			"The redirect_uri query parameter is required",
		)
		return
	}

	// Use first registered redirect URI if none provided
	if !ok || redirectURI == "" {
		redirectURI = client.RedirectUris[0]
	} else if !slices.Contains(client.RedirectUris, redirectURI) {
		errors.SendErrorResponse(c, http.StatusBadRequest, errors.InvalidRedirectURI, "The provided redirect_uri is not registered for this client")
		return
	}

	uri, parseErr := url.ParseRequestURI(redirectURI)
	if parseErr != nil {
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.InvalidRedirectURI,
			"The provided redirect_uri is not a valid URI",
		)
		return
	}

	state := c.Query("state")
	if state == "" {
		ctrl.redirectWithError(
			c,
			uri,
			"invalid_request",
			"The state query parameter is required",
			"",
		)
		return
	}
	if len(state) > 255 {
		ctrl.redirectWithError(
			c,
			uri,
			"invalid_request",
			"The state query parameter must not exceed 255 characters",
			state,
		)
		return
	}

	responseType, ok := c.GetQuery("response_type")
	if !ok || responseType == "" {
		ctrl.redirectWithError(
			c,
			uri,
			"invalid_request",
			"The response_type query parameter is required",
			state,
		)
		return
	}

	if responseType != "code" {
		ctrl.redirectWithError(
			c,
			uri,
			"unsupported_response_type",
			"The /oauth/authorize endpoint only supports the 'code' response type",
			state,
		)
		return
	}

	codeChallenge, ok := c.GetQuery("code_challenge")
	if !ok || codeChallenge == "" {
		ctrl.redirectWithError(
			c,
			uri,
			"invalid_request",
			"The code_challenge query parameter is required",
			state,
		)
		return
	}

	code, authErr := ctrl.service.Authorize(
		c.Request.Context(),
		client,
		codeChallenge,
		utils.User.Subject,
		c.ClientIP(),
	)
	if authErr != nil {
		ctrl.redirectWithError(c, uri, "server_error", "", state)
		return
	}

	q := uri.Query()
	q.Add("code", *code)
	q.Add("state", state)
	uri.RawQuery = q.Encode()

	c.Redirect(http.StatusFound, uri.String())
}

// Token handles the OAuth2 token endpoint.
// @Summary OAuth2 Token endpoint
// @Description Exchange authorization code for access token, refresh tokens, or use client credentials flow
// @Tags OAuth2
// @Accept application/x-www-form-urlencoded
// @Produce json
// @Security BasicAuth
// @Param grant_type formData string true "Grant type (authorization_code, client_credentials, or refresh_token)"
// @Param client_id formData string false "Client ID (required if not using Basic Auth)"
// @Param code formData string false "Authorization code (required for authorization_code grant)"
// @Param code_verifier formData string false "PKCE code verifier (required for authorization_code grant)"
// @Param refresh_token formData string false "Refresh token (required for refresh_token grant)"
// @Success 200 {object} TokenResponse "Token response with access token and optional refresh token"
// @Failure 400 {object} errors.APIError "Invalid request parameters or grant type"
// @Failure 401 {object} errors.APIError "Invalid client credentials"
// @Failure 500 {object} errors.APIError "Internal server error"
// @Router /oauth/token [post].
func (ctrl *Controller) Token(c *gin.Context) {
	_, errs := endpoint.SetupEndpoint[any](c, endpoint.WithoutBody())
	if len(errs) > 0 {
		endpoint.SendSetupErrorResponse(c, errs)
		return
	}

	contentType := c.GetHeader("Content-Type")
	if contentType != "application/x-www-form-urlencoded" {
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.InvalidContentType,
			"The Content-Type header must be application/x-www-form-urlencoded",
		)
		return
	}

	if err := c.Request.ParseForm(); err != nil {
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.InvalidRequestBody,
			"Failed to parse request body",
		)
		return
	}

	grantType := c.Request.FormValue("grant_type")
	if grantType == "" {
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.MissingGrantType,
			"The grant_type parameter is required",
		)
		return
	}
	var clientSecret = ""
	clientID := c.Request.FormValue("client_id")
	if clientID == "" {
		// Try to get client_id and secret from Basic Auth
		var ok bool
		clientID, clientSecret, ok = c.Request.BasicAuth()
		if !ok {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.MissingClientID,
				"Basic auth header is required if client_id is not provided in the body or request is confidential",
			)
		}
		if clientID == "" {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.MissingClientID,
				"The client_id parameter is required",
			)
			return
		}
	}

	client, err := ctrl.service.GetClient(c.Request.Context(), clientID, c.ClientIP())
	if err != nil {
		c.JSON(err.Code, err)
		return
	}

	if client.ClientSecretHash.Valid {
		if clientSecret == "" {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.MissingClientSecret,
				"Client Secret is required for confidential clients",
			)
			return
		}

		if !client.ClientSecretHash.Valid ||
			!tokens.CompareClientSecretHash(clientSecret, client.ClientSecretHash.String) {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.InvalidClientSecret,
				"Invalid Client Secret",
			)
			return
		}
	}

	switch grantType {
	case "authorization_code":
		if !slices.Contains(client.GrantTypes, database.GrantTypesAuthorizationCode) {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.InvalidGrantType,
				"The client is not authorized to use the authorization_code grant type",
			)
		}

		code := c.Request.FormValue("code")
		if code == "" {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.MissingCode,
				"The code parameter is required",
			)
			return
		}
		codeVerifier := c.Request.FormValue("code_verifier")
		if codeVerifier == "" {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.MissingCodeVerifier,
				"The code_verifier parameter is required",
			)
			return
		}

		accessToken, refreshToken, scopes, err := ctrl.service.AuthorizationCodeFlow(
			c.Request.Context(),
			client,
			code,
			codeVerifier,
			c.ClientIP(),
		)
		if err != nil {
			c.JSON(err.Code, err)
			return
		}

		tokenRes := TokenResponse{
			AccessToken:           *accessToken,
			AccessTokenExpiresIn:  int(client.AccessTokenValidDuration),
			RefreshTokenExpiresIn: int(client.RefreshTokenValidDuration),
			Scopes:                scopes,
		}

		if slices.Contains(client.GrantTypes, database.GrantTypesRefreshToken) {
			tokenRes.RefreshToken = *refreshToken
		}

		c.JSON(http.StatusOK, tokenRes)

	case "client_credentials":
		if !slices.Contains(client.GrantTypes, database.GrantTypesClientCredentials) {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.InvalidGrantType,
				"The client is not authorized to use the client_credentials grant type",
			)
		}

		accessToken, scopes, err := ctrl.service.ClientCredentialsFlow(
			client,
			c.ClientIP(),
		)
		if err != nil {
			c.JSON(err.Code, err)
			return
		}

		c.JSON(http.StatusOK, TokenResponse{
			AccessToken:          *accessToken,
			AccessTokenExpiresIn: int(client.AccessTokenValidDuration),
			Scopes:               scopes,
		})
	case "refresh_token":
		if !slices.Contains(client.GrantTypes, database.GrantTypesRefreshToken) {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.InvalidGrantType,
				"The client is not authorized to use the refresh_token grant type",
			)
		}

		refreshToken := c.Request.FormValue("refresh_token")
		if refreshToken == "" {
			errors.SendErrorResponse(
				c,
				http.StatusBadRequest,
				errors.MissingRefreshToken,
				"The refresh_token parameter is required",
			)
			return
		}

		newAccessToken, newRefreshToken, scopes, err := ctrl.service.RefreshTokenFlow(
			c.Request.Context(),
			client,
			refreshToken,
			c.ClientIP(),
		)
		if err != nil {
			c.JSON(err.Code, err)
			return
		}

		c.JSON(http.StatusOK, TokenResponse{
			AccessToken:           *newAccessToken,
			AccessTokenExpiresIn:  int(client.AccessTokenValidDuration),
			RefreshToken:          *newRefreshToken,
			RefreshTokenExpiresIn: int(client.RefreshTokenValidDuration),
			Scopes:                scopes,
		})

	default:
		errors.SendErrorResponse(
			c,
			http.StatusBadRequest,
			errors.InvalidGrantType,
			"The grant_type is not supported",
		)
	}
}

// redirectWithError redirects the user with OAuth2 error parameters.
func (ctrl *Controller) redirectWithError(
	c *gin.Context,
	redirectURI *url.URL,
	errorCode, errorDescription, state string,
) {
	q := redirectURI.Query()
	q.Add("error", errorCode)
	if errorDescription != "" {
		q.Add("error_description", errorDescription)
	}
	if state != "" {
		q.Add("state", state)
	}
	redirectURI.RawQuery = q.Encode()
	c.Redirect(http.StatusFound, redirectURI.String())
}